Remaining interactive commands (#792)
* Set contract build directory in contract schema

* Implement method to take contact methods from AST

* Implement method to get contract methods from AST
This method filters the contract methods (and its parents' methods) and
also group them by those that have the initializer modifier and those that
doesn't have it.

* Look for contracts in both local and package dirs
- Look for contracts in local build directory
- Look for contracts from dependencies
- Implement a general refactor

* Make the create command interactive

* Enhance Truffle.ts#getContracts method

* Remove .only in test

* Simplify create command functions

* Set session command interactive

* Make remove command interactive

* Make publish command interactive

* Improve initializer function check

* Enhance push command interactivity

* Use ZosPackageFile instance method to get deps

* Improve prompt#contractsList function
Now, we can get not only the contracts inside `build/contracts` but also
the contracts added in the zos.json file (this is particularly useful
for the remove command)

* Implement method to get list of proxies

Refactor function to get contract methods

Refactor in function to get contract method args

Add comments explaining prompt.ts functions

Rename functions in prompt.ts

Refactor tests

Fix import in ZosPackageFile

Implement update command interaction
 - Add an additional prompt question for asking the user which contract they want to update: by name, by address or all
 - Copied logic to prompt for init parameters from commands/create.ts to commands/update.ts

Define more accurate types

Cleaned up and simplified code

Refactor tests

Send message, name and type as parameters for contractsList and others

Add non-interactive flag to all commands

Call preceding commands if needed

Create singleton to persist compilation state

Implement normalize function to normalize promptFor output

Verify if contract exists before getting contract methods

* Add --no-interactive flag to integratiion tests

* Refactor commands files
- Standarized naming for variables (prompted, args, props, etc)
- Unified props into a function (setCommandProps)
- Moved other commands calls into a actionsWrapper function

* Simplified prompt.ts functions and fixed tests

* Fix input.ts tests

* Check if initMethod exists before creating a proxy
So the initializer function is not called if not needed

* Fix some linter errors

* More refactors

* Refactored all command files
- Standarized how opts, args, defaults and props are named
- Added remaining types in functions and parameters
- Created a setCommandProps function that wraps all inquirer properties
for each arg and opt

* Enhance tests

* Send name and message to networksList function

* Change some comments in prompt.ts

* Change condition statement on upgrade (ProxyAdmin)
Fixed the condition statement to decide whether to upgrade a proxy or to
upgrade it and also run a function afterwards.

* Run link command on create if needed

* Update packages/cli/src/commands/session.ts


* Apply @spalladino's suggestions from code review


* Implement some @spalladino suggestions

+++ packages/cli/src/models/compiler/Compiler.ts
@@ -3,27 +3,32 @@ import { FileSystem, Logger } from 'zos-lib';
 
 const log = new Logger('Compiler');
 
+const state = { alreadyCompiled: false };
+
 const Compiler = {
-  async call(): Promise<{ stdout: string, stderr: string }> {
-    log.info('Compiling contracts with Truffle...');
-    let truffleBin = `${process.cwd()}/node_modules/.bin/truffle`;
-    if (!FileSystem.exists(truffleBin)) truffleBin = 'truffle'; // Attempt to load global truffle if local was not found
+  async call(force: boolean = false): Promise<{ stdout: string, stderr: string }> {
+    if (force || !state.alreadyCompiled) {
+      log.info('Compiling contracts with Truffle...');
+      let truffleBin = `${process.cwd()}/node_modules/.bin/truffle`;
+      if (!FileSystem.exists(truffleBin)) truffleBin = 'truffle'; // Attempt to load global truffle if local was not found
 
-    return new Promise((resolve, reject) => {
-      const args: object = { shell: true };
+      return new Promise((resolve, reject) => {
+        const args: object = { shell: true };
 
-      execFile(truffleBin, ['compile', '--all'], args, (error: ExecException, stdout, stderr) => {
+        execFile(truffleBin, ['compile', '--all'], args, (error: ExecException, stdout, stderr) => {
 
-        if (error) {
-          if (error.code === 127) console.error('Could not find truffle executable. Please install it by running: npm install truffle');
-          reject(error);
-        } else {
-          resolve({ stdout, stderr });
-        }
-        if (stdout) console.log(stdout);
-        if (stderr) console.error(stderr);
+          if (error) {
+            if (error.code === 127) console.error('Could not find truffle executable. Please install it by running: npm install truffle');
+            reject(error);
+          } else {
+            state.alreadyCompiled = true;
+            resolve({ stdout, stderr });
+          }
+          if (stdout) console.log(stdout);
+          if (stderr) console.error(stderr);
+        });
       });
-    });
+    }
   }
 };
